% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggveganextra.R
\name{ordi_identify}
\alias{ordi_identify}
\title{Interactive identify}
\usage{
ordi_identify(plotname, ordiname, display)
}
\arguments{
\item{plotname}{Name of a plot created with ggvegan autoplot}

\item{ordiname}{Result of ordination PCA, RDA, CA, CCA from rda or cca}

\item{display}{What to label. Currently only accepts "species"}
}
\value{
The original ordiname is modified with labels
}
\description{
Interactive identify ggvegan species
}
\details{
This function is designed to be run interactively. First create a standard
ordination using rda() or cca(). Use autoplot to display the results.
}
\examples{
data(dune)
dune_pca <- rda(dune)
dune_plt <- autoplot(dune_pca, layers="species", geom="point", legend.position="none", arrows=FALSE)
dune_plt

ordi_identify(dune_plt, dune_pca, "species")
}
\author{
Roy Sanderson, School of Natural & Environmental Science, Newcastle
University roy.sanderson@newcastle.ac.uk
}
